tests/wpt/web-platform-tests/css/css-ui/historical/user-modify-01.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>user-modify should be unsupported in author origin</title>
<link rel="help" href="https://developer.mozilla.org/en-US/docs/Web/CSS/user-modify">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
#modified {
    user-modify: read-write;
}
</style>
<div id="inherited"></div>
<div id="modified"></div>
<script>
    test(function() {
        for (const el of [inherited, modified]) {
            assert_equals(el.style.userModify, undefined);
            assert_equals(el.style.UserModify, undefined);
            assert_equals(getComputedStyle(el).getPropertyValue('user-modify'), '');
        }
    }, 'user-modify is not supported');
</script>
